refactor(ast): move pure non-node helpers to @kubb/ast/utils
Relocate objectKey, buildObject, buildList, childName, enumPropName,
extractRefName and findDiscriminator from the main @kubb/ast barrel to the
@kubb/ast/utils subpath. These are pure helpers whose signatures involve no AST
nodes, so they belong with the source/string utilities rather than the node
tree. Helpers that take or return nodes stay on the main barrel.

adapter-oas now imports the OAS-resolution helpers from @kubb/ast/utils instead
of the @kubb/core ast namespace. Breaking: these names leave the main barrel.
