+++ packages/adapter-oas/package.json
@@ -56,6 +56,7 @@
     "typecheck": "tsc -p ./tsconfig.json --noEmit --emitDeclarationOnly false"
   },
   "dependencies": {
+    "@kubb/ast": "workspace:*",
     "@kubb/core": "workspace:*",
     "@redocly/openapi-core": "^2.31.6",
     "oas": "^32.1.18",
