+++ packages/ast/src/index.ts
@@ -31,14 +31,11 @@ export {
 } from './factory.ts'
 export { isHttpOperationNode, isInputNode, isOperationNode, isOutputNode, isSchemaNode, narrowSchema } from './guards.ts'
 export { createPrinterFactory, definePrinter } from './printer.ts'
-export { extractRefName } from './refs.ts'
-export { childName, collectImports, enumPropName, findDiscriminator } from './resolvers.ts'
+export { collectImports } from './resolvers.ts'
 export { schemaSignature } from './signature.ts'
 export { mergeAdjacentObjectsLazy, setDiscriminatorEnum, setEnumName, simplifyUnion } from './transformers.ts'
 export type * from './types.ts'
 export {
-  buildList,
-  buildObject,
   caseParams,
   collectUsedSchemaNames,
   containsCircularRef,
@@ -47,7 +44,6 @@ export {
   extractStringsFromNodes,
   findCircularSchemas,
   isStringType,
-  objectKey,
   syncSchemaRef,
 } from './utils/ast.ts'
 export { collect, transform, walk } from './visitor.ts'
