+++ packages/ast/src/utils/ast.test.ts
@@ -3,8 +3,6 @@ import { createFunctionParameter, createOperation, createParameter, createParams
 import type { OperationNode, ParameterNode } from '../types.ts'
 import type { OperationParamsResolver } from './ast.ts'
 import {
-  buildList,
-  buildObject,
   caseParams,
   collectReferencedSchemaNames,
   collectUsedSchemaNames,
@@ -13,7 +11,6 @@ import {
   createOperationParams,
   findCircularSchemas,
   isStringType,
-  objectKey,
   resolveRefName,
   syncSchemaRef,
 } from './ast.ts'
@@ -2085,74 +2082,5 @@ describe('collectUsedSchemaNames', () => {
     const result = collectUsedSchemaNames([createItemOp], [bodySchema])
 
     expect(result).toStrictEqual(new Set(['CreateItemBody']))
-  })
-})
-
-describe('buildList', () => {
-  it('returns an empty list for no items', () => {
-    expect(buildList([])).toMatchInlineSnapshot(`"[]"`)
-  })
-
-  it('keeps single-line items inline', () => {
-    expect(buildList(['z.string()', 'z.number()'])).toMatchInlineSnapshot(`"[z.string(), z.number()]"`)
-  })
-
-  it('wraps and indents when an item spans multiple lines', () => {
-    const member = buildObject(['id: z.number()'])
-    expect(buildList([`z.object(${member})`, 'z.string()'])).toMatchInlineSnapshot(`
-      "[
-        z.object({
-          id: z.number(),
-        }),
-        z.string(),
-      ]"
-    `)
-  })
-
-  it('uses custom brackets', () => {
-    expect(buildList(['a', 'b'], ['(', ')'])).toMatchInlineSnapshot(`"(a, b)"`)
-  })
-})
-
-describe('objectKey', () => {
-  it('leaves valid identifiers unquoted', () => {
-    expect(objectKey('id')).toMatchInlineSnapshot(`"id"`)
-  })
-
-  it('leaves reserved words and globals unquoted', () => {
-    expect(objectKey('name')).toMatchInlineSnapshot(`"name"`)
-    expect(objectKey('class')).toMatchInlineSnapshot(`"class"`)
-  })
-
-  it('single-quotes keys that are not valid identifiers', () => {
-    expect(objectKey('x-total')).toMatchInlineSnapshot(`"'x-total'"`)
-    expect(objectKey('200')).toMatchInlineSnapshot(`"'200'"`)
-  })
-})
-
-describe('buildObject', () => {
-  it('returns an empty object literal for no entries', () => {
-    expect(buildObject([])).toMatchInlineSnapshot(`"{}"`)
-  })
-
-  it('indents entries and adds a trailing comma', () => {
-    expect(buildObject(['id: z.number()', 'name: z.string()'])).toMatchInlineSnapshot(`
-      "{
-        id: z.number(),
-        name: z.string(),
-      }"
-    `)
-  })
-
-  it('indents a nested object cumulatively', () => {
-    const address = `address: ${buildObject(['street: z.string()'])}`
-    expect(buildObject(['id: z.number()', address])).toMatchInlineSnapshot(`
-      "{
-        id: z.number(),
-        address: {
-          street: z.string(),
-        },
-      }"
-    `)
   })
 })
